chore: refactor & removed code duplicate
