--- apps/backoffice-v2/src/common/components/molecules/ProcessTracker/trackers/collection-flow/components/CollectionFlowProcessTitle/components/RequestProcesses/hooks/useRequestProcessDialog/useRequestProcessDialog.ts
@@ -1,10 +1,0 @@
-import { useState } from 'react';
-
-export const useRequestProcessDialog = () => {
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-
-  return {
-    isDialogOpen,
-    onOpenChange: setIsDialogOpen,
-  };
-};
