--- apps/backoffice-v2/src/common/components/molecules/ProcessTracker/trackers/collection-flow/components/CollectionFlowProcessTitle/components/RequestProcesses/index.ts
@@ -1,1 +1,0 @@
-export * from './RequestProcesses';
