--- apps/backoffice-v2/src/common/components/molecules/ProcessTracker/trackers/collection-flow/components/CollectionFlowProcessTitle/index.ts
@@ -1,1 +1,0 @@
-export * from './CollectionFlowProcessTitle';
