+++ apps/backoffice-v2/src/common/components/molecules/ProcessTracker/trackers/collection-flow/components/CollectionFlowStepItem/components/CollectionFlowProcessTitle/CollectionFlowProcessTitle.tsx
@@ -1,14 +1,12 @@
 import { ctw } from '@/common/utils/ctw/ctw';
-import { TWorkflowById } from '@/domains/workflows/fetchers';
-import { useIsWorkflowStepsCanBeRevised } from '../../hooks/useIsWorkflowStepsCanBeRevised';
 import { RequestProcesses } from './components/RequestProcesses';
 import { useStepsRequesting } from './hooks/useStepsRequesting';
+import { useTracker } from '../../../../../../components/Tracker/hooks/useTracker';
+import { Tracker } from '../../../../../../components/Tracker/Tracker';
+import { useIsWorkflowStepsCanBeRevised } from '../../hooks/useIsWorkflowStepsCanBeRevised';
 
-interface ICollectionFlowProcessTitleProps {
-  workflow: TWorkflowById;
-}
-
-export const CollectionFlowProcessTitle = ({ workflow }: ICollectionFlowProcessTitleProps) => {
+export const CollectionFlowProcessTitle = () => {
+  const { workflow } = useTracker();
   const { stepsCountToRequest, isLoading, sendRequestedStepsToRevision } =
     useStepsRequesting(workflow);
   const isShouldDisplayRequestButton = stepsCountToRequest > 0;
@@ -21,14 +19,14 @@ export const CollectionFlowProcessTitle = ({ workflow }: ICollectionFlowProcessT
 
   return (
     <div className="flex w-full flex-row items-center justify-between gap-2 pr-2 !no-underline hover:no-underline">
-      <div
-        className={ctw('whitespace-nowrap no-underline', {
+      <Tracker.Title
+        className={ctw({
           ['max-w-[60px] overflow-hidden text-ellipsis']: isShouldDisplayRequestButton,
         })}
         title={isShouldDisplayRequestButton ? 'Collection Flow' : undefined}
       >
         Collection Flow
-      </div>
+      </Tracker.Title>
       {stepsCountToRequest > 0 && (
         <RequestProcesses
           requestCount={stepsCountToRequest}
